Update UI Automation representation of the grid and its items.
Incorporate changes from Fons Sonnemans which fix UIA hierarchy for grid and its items.

Change grid and items' UIA Name, ControlType, and LocalizedControlTypes to be a more logical match for the elements, and remove duplication on content between these properties.

Had GridView support the UIA Grid pattern, and items support the UIA GridItem pattern.

Only expose item row and column data through the UIA GridItem pattern, as any screen reader can access that when necessary.

Raise a UIA Notification event so that screen reader announce something helpful when a cell is invoked.

Change unnecessary SelectionMode of Single, to None. A side effect of this is that cells are now invokable via Narrator touch as required.